churn-ga-xgb/requirements.txt
# Core ML libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Model interpretation
shap>=0.40.0

# Model persistence
joblib>=1.1.0

# Additional utilities
tqdm>=4.62.0
psutil>=5.8.0

# Optional: For better performance
numba>=0.56.0

# Development and testing (optional)
pytest>=6.0.0
black>=21.0.0
flake8>=3.9.0

pandas>=1.3.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
